@@ -26,32 +26,23 @@
 import Foundation
 
 /// A reusable swift error struct
-public struct ErrorMessage: LocalizedError, Equatable, Hashable, CustomStringConvertible, Sendable {
+public struct ErrorMessage: Error, Equatable, Hashable, CustomStringConvertible, Sendable {
 
     // MARK: - Initializers
 
     /// Create an `ErrorMessage`
     /// - Parameters:
     ///   - message: The error message
-    ///   - failureReason: A description of the failure that caused the error to occur
-    ///   - recoverySuggestion: A suggestion for how to recover from the error
-    ///   - helpAnchor: The help anchor
     ///   - file: File containing the callsite where the error occured
     ///   - function: Function containing the callsite where the error occured
     ///   - line: Line number of the callsite where the error occured
     ///   - column: Column number of the callsite where the error occured
     public init(_ message: String,
-                failureReason: String? = nil,
-                recoverySuggestion: String? = nil,
-                helpAnchor: String? = nil,
                 file: StaticString = #file,
                 function: StaticString = #function,
                 line: UInt = #line,
                 column: UInt = #column) {
         self.message = message
-        self.failureReason = failureReason
-        self.recoverySuggestion = recoverySuggestion
-        self.helpAnchor = helpAnchor
         self.file = file
         self.function = function
         self.line = line
@@ -78,36 +69,18 @@ public struct ErrorMessage: LocalizedError, Equatable, Hashable, CustomStringCon
     // MARK: - CustomStringConvertible
 
     public var description: String {
-        [
-            [file.description,
-             function.description,
-             line.description,
-             column.description, message]
-                .joined(separator: ":"),
-            failureReason,
-            recoverySuggestion,
-            helpAnchor
-        ]
-        .compactMap { $0 }
-        .joined(separator: ", ")
+        [file.description,
+         function.description,
+         line.description,
+         column.description,
+         message]
+            .joined(separator: ":")
     }
-
-    // MARK: - LocalizedError
-
-    public var errorDescription: String? {
-        message
-    }
-
-    public let failureReason: String?
-
-    public let recoverySuggestion: String?
-
-    public let helpAnchor: String?
 
     // MARK: - Equatable
 
     public static func == (lhs: ErrorMessage, rhs: ErrorMessage) -> Bool {
-        lhs.message == rhs.message && lhs.file.description == rhs.file.description && lhs.function.description == rhs.function.description && lhs.line == rhs.line && lhs.column == rhs.column && lhs.failureReason == rhs.failureReason && lhs.recoverySuggestion == rhs.recoverySuggestion && lhs.helpAnchor == rhs.helpAnchor
+        lhs.message == rhs.message && lhs.file.description == rhs.file.description && lhs.function.description == rhs.function.description && lhs.line == rhs.line && lhs.column == rhs.column
     }
 
     // MARK: - Hashable
@@ -118,8 +91,5 @@ public struct ErrorMessage: LocalizedError, Equatable, Hashable, CustomStringCon
         hasher.combine(function.description)
         hasher.combine(line)
         hasher.combine(column)
-        hasher.combine(failureReason)
-        hasher.combine(recoverySuggestion)
-        hasher.combine(helpAnchor)
     }
 }
